+++ jabiseo-api/src/main/java/com/jabiseo/auth/application/oidc/OauthMemberInfo.java
@@ -1,4 +1,4 @@
-package com.jabiseo.auth.oidc;
+package com.jabiseo.auth.application.oidc;
 
 import lombok.Builder;
 import lombok.Getter;
