+++ jabiseo-api/src/main/java/com/jabiseo/auth/application/oidc/TokenValidatorManager.java
@@ -1,4 +1,4 @@
-package com.jabiseo.auth.oidc;
+package com.jabiseo.auth.application.oidc;
 
 
 import com.jabiseo.auth.exception.AuthenticationBusinessException;
