+++ jabiseo-api/src/main/java/com/jabiseo/auth/application/oidc/property/OidcIdTokenProperty.java
@@ -1,0 +1,13 @@
+package com.jabiseo.auth.application.oidc.property;
+
+
+/**
+ * @param issuer   발행 회사 url
+ * @param audience client-id
+ */
+public record OidcIdTokenProperty(
+        String issuer,
+        String audience
+) {
+
+}
