+++ jabiseo-api/src/main/java/com/jabiseo/auth/application/usecase/LogoutUseCase.java
@@ -1,10 +1,9 @@
-package com.jabiseo.auth.usecase;
+package com.jabiseo.auth.application.usecase;
 
 import org.springframework.stereotype.Service;
 
 @Service
-public class WithdrawUseCase {
-
+public class LogoutUseCase {
     public void execute() {
     }
