+++ jabiseo-api/src/test/java/com/jabiseo/auth/oidc/KakaoIdTokenValidatorTest.java
@@ -1,10 +1,9 @@
 package com.jabiseo.auth.oidc;
 
+import com.jabiseo.auth.application.oidc.KakaoIdTokenValidator;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 
 @SpringBootTest
