+++ jabiseo-domain/src/main/java/com/jabiseo/auth/domain/Token.java
@@ -1,0 +1,7 @@
+package com.jabiseo.auth.domain;
+
+
+public class Token {
+
+    private String value;
+}
