+++ jabiseo-infrastructure/build.gradle
@@ -10,6 +10,7 @@ jar {
 dependencies {
 
     implementation 'org.springframework.boot:spring-boot-starter-data-redis'
+    implementation 'org.springframework.boot:spring-boot-starter-web'
 
     runtimeOnly 'com.mysql:mysql-connector-j'
