+++ jabiseo-infrastructure/src/main/java/com/jabiseo/cache/CacheRepository.java
@@ -1,0 +1,16 @@
+package com.jabiseo.cache;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import org.springframework.data.redis.core.RedisTemplate;
+import org.springframework.data.redis.core.ValueOperations;
+import org.springframework.stereotype.Component;
+
+@Component
+public class CacheRepository {
+
+
+    public void save(String key, String value, int ttl){
+
+    }
+
+}
