+++ jabiseo-infrastructure/src/main/java/com/jabiseo/client/KakaoKauthClient.java
@@ -1,9 +1,10 @@
-package com.jabiseo.auth.oidc;
+package com.jabiseo.client;
 
-
+;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.service.annotation.GetExchange;
 import org.springframework.web.service.annotation.HttpExchange;
+
 
 
 @HttpExchange
