+++ jabiseo-infrastructure/src/main/java/com/jabiseo/client/RestClientConfig.java
@@ -1,6 +1,6 @@
-package com.jabiseo.config;
+package com.jabiseo.client;
 
-import com.jabiseo.auth.oidc.KakaoKauthClient;
+
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.client.RestClient;
